Dragon install: probe the registry and detect multiple Dragon versions, use environment variables to look up install path. re #3.
Do not always assume that C is the sysdrive letter - it might be D, E, X, etc. Thus use programfiles environment variable to locate Dragon installation path.
Also, probe the registry to find installed versions of Dragon (not just 15). The next commit will introduce this ability, as well as a single choice dialog to ask which version the custom commands shuold be applied to (if there's only one, just use the first).